move home to pages dir and update imports
