+++ src/index.js
@@ -1,3 +1,3 @@
-import { loadHome } from "./home"
+import { loadHome } from "./pages/home"
 import maincss from "./stylesheets/styles.css"
 loadHome()
